backend/sources/edsm/requirements.txt
absl-py==0.10.0
apache-beam[gcp]==2.27.0
google-cloud-storage==1.31.2